+++ packages/react/runtime/__test__/list.test.jsx
@@ -1713,6 +1713,7 @@ describe('list reload', () => {
                     "full-span": false,
                     "item-key": "2",
                     "to": 1,
+                    "type": "__Card__:__snapshot_a94a8_test_39",
                   },
                 ],
               },
@@ -1951,9 +1952,9 @@ describe('list-item JSXSpread', () => {
       b.insertBefore(d2);
       b.insertBefore(d3);
 
-      d1.setAttribute(0, { 'item-key': '1', 'full-span': true });
-      d2.setAttribute(0, { 'item-key': '2', 'full-span': true });
-      d3.setAttribute(0, { 'item-key': '3', 'full-span': true });
+      d1.setAttribute(0, { 'item-key': '1', 'full-span': true, 'recyclable': true });
+      d2.setAttribute(0, { 'item-key': '2', 'full-span': true, 'recyclable': true });
+      d3.setAttribute(0, { 'item-key': '3', 'full-span': true, 'recyclable': true });
     }
 
     __pendingListUpdates.flush();
@@ -1967,9 +1968,9 @@ describe('list-item JSXSpread', () => {
     bb.insertBefore(d2);
     bb.insertBefore(d3);
 
-    d1.setAttribute(0, { 'item-key': '1', 'full-span': true });
-    d2.setAttribute(0, { 'item-key': '2', 'full-span': false });
-    d3.setAttribute(0, { 'item-key': '3', 'full-span': true });
+    d1.setAttribute(0, { 'item-key': '1', 'full-span': true, 'recyclable': false });
+    d2.setAttribute(0, { 'item-key': '2', 'full-span': false, 'recyclable': false });
+    d3.setAttribute(0, { 'item-key': '3', 'full-span': true, 'recyclable': false });
 
     hydrate(b, bb);
     b.unRenderElements();
@@ -1991,18 +1992,21 @@ describe('list-item JSXSpread', () => {
                     "full-span": true,
                     "item-key": "1",
                     "position": 0,
+                    "recyclable": true,
                     "type": "__Card__:__snapshot_a94a8_test_46",
                   },
                   {
                     "full-span": true,
                     "item-key": "2",
                     "position": 1,
+                    "recyclable": true,
                     "type": "__Card__:__snapshot_a94a8_test_46",
                   },
                   {
                     "full-span": true,
                     "item-key": "3",
                     "position": 2,
+                    "recyclable": true,
                     "type": "__Card__:__snapshot_a94a8_test_46",
                   },
                 ],
@@ -2015,10 +2019,30 @@ describe('list-item JSXSpread', () => {
                 "updateAction": [
                   {
                     "flush": false,
+                    "from": 0,
+                    "full-span": true,
+                    "item-key": "1",
+                    "recyclable": false,
+                    "to": 0,
+                    "type": "__Card__:__snapshot_a94a8_test_46",
+                  },
+                  {
+                    "flush": false,
                     "from": 1,
                     "full-span": false,
                     "item-key": "2",
+                    "recyclable": false,
                     "to": 1,
+                    "type": "__Card__:__snapshot_a94a8_test_46",
+                  },
+                  {
+                    "flush": false,
+                    "from": 2,
+                    "full-span": true,
+                    "item-key": "3",
+                    "recyclable": false,
+                    "to": 2,
+                    "type": "__Card__:__snapshot_a94a8_test_46",
                   },
                 ],
               },
@@ -2047,18 +2071,21 @@ describe('list-item JSXSpread', () => {
                     "full-span": true,
                     "item-key": "1",
                     "position": 0,
+                    "recyclable": true,
                     "type": "__Card__:__snapshot_a94a8_test_46",
                   },
                   {
                     "full-span": true,
                     "item-key": "2",
                     "position": 1,
+                    "recyclable": true,
                     "type": "__Card__:__snapshot_a94a8_test_46",
                   },
                   {
                     "full-span": true,
                     "item-key": "3",
                     "position": 2,
+                    "recyclable": true,
                     "type": "__Card__:__snapshot_a94a8_test_46",
                   },
                 ],
@@ -2071,10 +2098,30 @@ describe('list-item JSXSpread', () => {
                 "updateAction": [
                   {
                     "flush": false,
+                    "from": 0,
+                    "full-span": true,
+                    "item-key": "1",
+                    "recyclable": false,
+                    "to": 0,
+                    "type": "__Card__:__snapshot_a94a8_test_46",
+                  },
+                  {
+                    "flush": false,
                     "from": 1,
                     "full-span": false,
                     "item-key": "2",
+                    "recyclable": false,
                     "to": 1,
+                    "type": "__Card__:__snapshot_a94a8_test_46",
+                  },
+                  {
+                    "flush": false,
+                    "from": 2,
+                    "full-span": true,
+                    "item-key": "3",
+                    "recyclable": false,
+                    "to": 2,
+                    "type": "__Card__:__snapshot_a94a8_test_46",
                   },
                 ],
               },
@@ -2114,18 +2161,21 @@ describe('list-item JSXSpread', () => {
                     "full-span": true,
                     "item-key": "1",
                     "position": 0,
+                    "recyclable": true,
                     "type": "__Card__:__snapshot_a94a8_test_46",
                   },
                   {
                     "full-span": true,
                     "item-key": "2",
                     "position": 1,
+                    "recyclable": true,
                     "type": "__Card__:__snapshot_a94a8_test_46",
                   },
                   {
                     "full-span": true,
                     "item-key": "3",
                     "position": 2,
+                    "recyclable": true,
                     "type": "__Card__:__snapshot_a94a8_test_46",
                   },
                 ],
@@ -2138,10 +2188,30 @@ describe('list-item JSXSpread', () => {
                 "updateAction": [
                   {
                     "flush": false,
+                    "from": 0,
+                    "full-span": true,
+                    "item-key": "1",
+                    "recyclable": false,
+                    "to": 0,
+                    "type": "__Card__:__snapshot_a94a8_test_46",
+                  },
+                  {
+                    "flush": false,
                     "from": 1,
                     "full-span": false,
                     "item-key": "2",
+                    "recyclable": false,
                     "to": 1,
+                    "type": "__Card__:__snapshot_a94a8_test_46",
+                  },
+                  {
+                    "flush": false,
+                    "from": 2,
+                    "full-span": true,
+                    "item-key": "3",
+                    "recyclable": false,
+                    "to": 2,
+                    "type": "__Card__:__snapshot_a94a8_test_46",
                   },
                 ],
               },
@@ -2161,6 +2231,253 @@ describe('list-item JSXSpread', () => {
           </list-item>
         </list>
       </view>
+    `);
+  });
+});
+
+describe('list-item with platform info attributes', () => {
+  const s1 = __SNAPSHOT__(
+    <view>
+      <text>111</text>
+      <list id='list'>{HOLE}</list>
+    </view>,
+  );
+
+  const s3 = __SNAPSHOT__(
+    <list-item
+      item-key={HOLE}
+      reuse-identifier={HOLE}
+      full-span={HOLE}
+      sticky-top={HOLE}
+      sticky-bottom={HOLE}
+      estimated-height={HOLE}
+      estimated-height-px={HOLE}
+      estimated-main-axis-size-px={HOLE}
+      recyclable={HOLE}
+    >
+      <text>World</text>
+    </list-item>,
+  );
+
+  it('basic list-item with platform info attributes', () => {
+    const b = new SnapshotInstance(s1);
+    b.ensureElements();
+    const root = b.__elements[0];
+    const listRef = b.__elements[3];
+
+    {
+      const d0 = new SnapshotInstance(s3);
+      const d1 = new SnapshotInstance(s3);
+      const d2 = new SnapshotInstance(s3);
+      b.insertBefore(d0);
+      b.insertBefore(d1);
+      b.insertBefore(d2);
+
+      d0.setAttribute(0, {
+        'item-key': 'list-item-0',
+        'reuse-identifier': 'A',
+        'full-span': true,
+        'sticky-top': true,
+        'sticky-bottom': false,
+        'estimated-height': 100,
+        'estimated-height-px': 100,
+        'estimated-main-axis-size-px': 100,
+        'recyclable': false,
+      });
+      d1.setAttribute(0, {
+        'item-key': 'list-item-1',
+        'reuse-identifier': 'A',
+        'full-span': false,
+        'sticky-top': false,
+        'sticky-bottom': false,
+        'estimated-height': 100,
+        'estimated-height-px': 100,
+        'estimated-main-axis-size-px': 100,
+        'recyclable': true,
+      });
+      d2.setAttribute(0, {
+        'item-key': 'list-item-2',
+        'reuse-identifier': 'A',
+        'full-span': true,
+        'sticky-top': false,
+        'sticky-bottom': true,
+        'estimated-height': 100,
+        'estimated-height-px': 100,
+        'estimated-main-axis-size-px': 100,
+        'recyclable': false,
+      });
+    }
+
+    __pendingListUpdates.flush();
+
+    expect(root).toMatchInlineSnapshot(`
+      <view>
+        <text>
+          <raw-text
+            text="111"
+          />
+        </text>
+        <list
+          id="list"
+          update-list-info={
+            [
+              {
+                "insertAction": [
+                  {
+                    "estimated-height": 100,
+                    "estimated-height-px": 100,
+                    "estimated-main-axis-size-px": 100,
+                    "full-span": true,
+                    "item-key": "list-item-0",
+                    "position": 0,
+                    "recyclable": false,
+                    "reuse-identifier": "A",
+                    "sticky-bottom": false,
+                    "sticky-top": true,
+                    "type": "__Card__:__snapshot_a94a8_test_48",
+                  },
+                  {
+                    "estimated-height": 100,
+                    "estimated-height-px": 100,
+                    "estimated-main-axis-size-px": 100,
+                    "full-span": false,
+                    "item-key": "list-item-1",
+                    "position": 1,
+                    "recyclable": true,
+                    "reuse-identifier": "A",
+                    "sticky-bottom": false,
+                    "sticky-top": false,
+                    "type": "__Card__:__snapshot_a94a8_test_48",
+                  },
+                  {
+                    "estimated-height": 100,
+                    "estimated-height-px": 100,
+                    "estimated-main-axis-size-px": 100,
+                    "full-span": true,
+                    "item-key": "list-item-2",
+                    "position": 2,
+                    "recyclable": false,
+                    "reuse-identifier": "A",
+                    "sticky-bottom": true,
+                    "sticky-top": false,
+                    "type": "__Card__:__snapshot_a94a8_test_48",
+                  },
+                ],
+                "removeAction": [],
+                "updateAction": [],
+              },
+            ]
+          }
+        />
+      </view>
+    `);
+
+    {
+      elementTree.triggerComponentAtIndex(listRef, 0);
+      elementTree.triggerComponentAtIndex(listRef, 1);
+      elementTree.triggerComponentAtIndex(listRef, 2);
+    }
+
+    // All virtual attributes: recyclable and reuse-identifier should not be set to list-item element.
+    expect(listRef).toMatchInlineSnapshot(`
+      <list
+        id="list"
+        update-list-info={
+          [
+            {
+              "insertAction": [
+                {
+                  "estimated-height": 100,
+                  "estimated-height-px": 100,
+                  "estimated-main-axis-size-px": 100,
+                  "full-span": true,
+                  "item-key": "list-item-0",
+                  "position": 0,
+                  "recyclable": false,
+                  "reuse-identifier": "A",
+                  "sticky-bottom": false,
+                  "sticky-top": true,
+                  "type": "__Card__:__snapshot_a94a8_test_48",
+                },
+                {
+                  "estimated-height": 100,
+                  "estimated-height-px": 100,
+                  "estimated-main-axis-size-px": 100,
+                  "full-span": false,
+                  "item-key": "list-item-1",
+                  "position": 1,
+                  "recyclable": true,
+                  "reuse-identifier": "A",
+                  "sticky-bottom": false,
+                  "sticky-top": false,
+                  "type": "__Card__:__snapshot_a94a8_test_48",
+                },
+                {
+                  "estimated-height": 100,
+                  "estimated-height-px": 100,
+                  "estimated-main-axis-size-px": 100,
+                  "full-span": true,
+                  "item-key": "list-item-2",
+                  "position": 2,
+                  "recyclable": false,
+                  "reuse-identifier": "A",
+                  "sticky-bottom": true,
+                  "sticky-top": false,
+                  "type": "__Card__:__snapshot_a94a8_test_48",
+                },
+              ],
+              "removeAction": [],
+              "updateAction": [],
+            },
+          ]
+        }
+      >
+        <list-item
+          estimated-height={100}
+          estimated-height-px={100}
+          estimated-main-axis-size-px={100}
+          full-span={true}
+          item-key="list-item-0"
+          sticky-bottom={false}
+          sticky-top={true}
+        >
+          <text>
+            <raw-text
+              text="World"
+            />
+          </text>
+        </list-item>
+        <list-item
+          estimated-height={100}
+          estimated-height-px={100}
+          estimated-main-axis-size-px={100}
+          full-span={false}
+          item-key="list-item-1"
+          sticky-bottom={false}
+          sticky-top={false}
+        >
+          <text>
+            <raw-text
+              text="World"
+            />
+          </text>
+        </list-item>
+        <list-item
+          estimated-height={100}
+          estimated-height-px={100}
+          estimated-main-axis-size-px={100}
+          full-span={true}
+          item-key="list-item-2"
+          sticky-bottom={true}
+          sticky-top={false}
+        >
+          <text>
+            <raw-text
+              text="World"
+            />
+          </text>
+        </list-item>
+      </list>
     `);
   });
 });
@@ -2528,7 +2845,7 @@ describe('list-item with "defer" attribute', () => {
                   {
                     "item-key": "1",
                     "position": 0,
-                    "type": "__Card__:__snapshot_a94a8_test_50",
+                    "type": "__Card__:__snapshot_a94a8_test_52",
                   },
                 ],
                 "removeAction": [],
@@ -2565,7 +2882,7 @@ describe('list-item with "defer" attribute', () => {
                   {
                     "item-key": "1",
                     "position": 0,
-                    "type": "__Card__:__snapshot_a94a8_test_50",
+                    "type": "__Card__:__snapshot_a94a8_test_52",
                   },
                 ],
                 "removeAction": [],
@@ -2658,17 +2975,17 @@ describe('list-item with "defer" attribute', () => {
                     {
                       "item-key": "0",
                       "position": 0,
-                      "type": "__Card__:__snapshot_a94a8_test_55",
+                      "type": "__Card__:__snapshot_a94a8_test_57",
                     },
                     {
                       "item-key": "1",
                       "position": 1,
-                      "type": "__Card__:__snapshot_a94a8_test_56",
+                      "type": "__Card__:__snapshot_a94a8_test_58",
                     },
                     {
                       "item-key": "2",
                       "position": 2,
-                      "type": "__Card__:__snapshot_a94a8_test_57",
+                      "type": "__Card__:__snapshot_a94a8_test_59",
                     },
                   ],
                   "removeAction": [],
@@ -2741,17 +3058,17 @@ describe('list-item with "defer" attribute', () => {
                     {
                       "item-key": "0",
                       "position": 0,
-                      "type": "__Card__:__snapshot_a94a8_test_55",
+                      "type": "__Card__:__snapshot_a94a8_test_57",
                     },
                     {
                       "item-key": "1",
                       "position": 1,
-                      "type": "__Card__:__snapshot_a94a8_test_56",
+                      "type": "__Card__:__snapshot_a94a8_test_58",
                     },
                     {
                       "item-key": "2",
                       "position": 2,
-                      "type": "__Card__:__snapshot_a94a8_test_57",
+                      "type": "__Card__:__snapshot_a94a8_test_59",
                     },
                   ],
                   "removeAction": [],
@@ -2822,17 +3139,17 @@ describe('list-item with "defer" attribute', () => {
                   {
                     "item-key": "0",
                     "position": 0,
-                    "type": "__Card__:__snapshot_a94a8_test_60",
+                    "type": "__Card__:__snapshot_a94a8_test_62",
                   },
                   {
                     "item-key": "1",
                     "position": 1,
-                    "type": "__Card__:__snapshot_a94a8_test_60",
+                    "type": "__Card__:__snapshot_a94a8_test_62",
                   },
                   {
                     "item-key": "2",
                     "position": 2,
-                    "type": "__Card__:__snapshot_a94a8_test_60",
+                    "type": "__Card__:__snapshot_a94a8_test_62",
                   },
                 ],
                 "removeAction": [],
@@ -2869,17 +3186,17 @@ describe('list-item with "defer" attribute', () => {
                   {
                     "item-key": "0",
                     "position": 0,
-                    "type": "__Card__:__snapshot_a94a8_test_60",
+                    "type": "__Card__:__snapshot_a94a8_test_62",
                   },
                   {
                     "item-key": "1",
                     "position": 1,
-                    "type": "__Card__:__snapshot_a94a8_test_60",
+                    "type": "__Card__:__snapshot_a94a8_test_62",
                   },
                   {
                     "item-key": "2",
                     "position": 2,
-                    "type": "__Card__:__snapshot_a94a8_test_60",
+                    "type": "__Card__:__snapshot_a94a8_test_62",
                   },
                 ],
                 "removeAction": [],
@@ -2928,17 +3245,17 @@ describe('list-item with "defer" attribute', () => {
                     {
                       "item-key": "0",
                       "position": 0,
-                      "type": "__Card__:__snapshot_a94a8_test_60",
+                      "type": "__Card__:__snapshot_a94a8_test_62",
                     },
                     {
                       "item-key": "1",
                       "position": 1,
-                      "type": "__Card__:__snapshot_a94a8_test_60",
+                      "type": "__Card__:__snapshot_a94a8_test_62",
                     },
                     {
                       "item-key": "2",
                       "position": 2,
-                      "type": "__Card__:__snapshot_a94a8_test_60",
+                      "type": "__Card__:__snapshot_a94a8_test_62",
                     },
                   ],
                   "removeAction": [],