+++ packages/react/runtime/src/internal.ts
@@ -23,7 +23,7 @@ export const __DynamicPartSlot: DynamicPartType = DynamicPartType.Slot;
 export const __DynamicPartMultiChildren: DynamicPartType = DynamicPartType.MultiChildren;
 export const __DynamicPartChildren: DynamicPartType = DynamicPartType.Children;
 export const __DynamicPartListChildren: DynamicPartType = DynamicPartType.ListChildren;
-export const __DynamicPartChildren_0: [DynamicPartType, number][] = [[DynamicPartType.Children, 0]];
+export { __DynamicPartChildren_0 } from './snapshot.js';
 
 export { updateSpread } from './snapshot/spread.js';
 export { updateEvent } from './snapshot/event.js';