+++ packages/rspeedy/create-rspeedy/template-react-vitest-rltl-ts/src/__tests__/index.test.tsx
@@ -12,7 +12,7 @@ test('App', async () => {
 
   render(
     <App
-      onMounted={() => {
+      onRender={() => {
         cb(`__MAIN_THREAD__: ${__MAIN_THREAD__}`)
       }}
     />,