+++ packages/web-platform/inline-style-parser/src/utils.rs
@@ -1,4 +1,4 @@
-use crate::*;
+use crate::char_code_definitions::*;
 
 pub fn cmp_str(test_str: &[u16], start: usize, end: usize, reference_str: &[u16]) -> bool {
   if (end - start) != reference_str.len() {
@@ -15,7 +15,7 @@ pub fn cmp_str(test_str: &[u16], start: usize, end: usize, reference_str: &[u16]
     let reference_code = reference_str[reference_str_offset];
     let mut test_code = test_str[i];
     // testCode.toLowerCase() for A..Z
-    if is_uppercase_letter!(test_code) {
+    if is_uppercase_letter(test_code) {
       test_code |= 32;
     }
     if test_code != reference_code {
@@ -29,7 +29,7 @@ pub fn find_white_space_end(source: &[u16], offset: usize) -> usize {
   let mut offset = offset;
   while offset < source.len() {
     let code = source[offset];
-    if !is_white_space!(code) {
+    if !is_white_space(code) {
       break;
     }
     offset += 1;
@@ -41,7 +41,7 @@ pub fn find_decimal_number_end(source: &[u16], offset: usize) -> usize {
   let mut offset = offset;
   while offset < source.len() {
     let code = source[offset];
-    if !is_digit!(code) {
+    if !is_digit(code) {
       break;
     }
     offset += 1;
@@ -56,20 +56,20 @@ pub fn consume_escaped(source: &[u16], offset: usize) -> usize {
   let mut offset = offset + 2;
   let source_length = source.len();
   // hex digit
-  if is_hex_digit!(get_char_code!(source, source_length, offset - 1)) {
+  if is_hex_digit(get_char_code(source, source_length, offset - 1)) {
     // It assumes that the U+005C REVERSE SOLIDUS (\) has already been consumed and
     // that the next input code point has already been verified to be part of a valid escape.
     let max_offset = core::cmp::min(offset + 5, source_length);
     while offset < max_offset {
-      if !is_hex_digit!(get_char_code!(source, source_length, offset)) {
+      if !is_hex_digit(get_char_code(source, source_length, offset)) {
         break;
       }
       offset += 1;
     }
     // If the next input code point is whitespace, consume it as well.
-    let code = get_char_code!(source, source_length, offset);
-    if is_white_space!(code) {
-      offset += get_new_line_length!(source, source_length, offset, code);
+    let code = get_char_code(source, source_length, offset);
+    if is_white_space(code) {
+      offset += get_new_line_length(source, source_length, offset, code);
     }
   }
 
@@ -85,14 +85,14 @@ pub fn consume_name(source: &[u16], offset: usize) -> usize {
   let source_length = source.len();
   while offset < source_length {
     let code = source[offset];
-    if is_name!(code) {
+    if is_name(code) {
       // Append the code point to result.
       offset += 1;
       continue;
     }
 
     // the stream starts with a valid escape
-    if is_valid_escape!(code, get_char_code!(source, source_length, offset + 1)) {
+    if is_valid_escape(code, get_char_code(source, source_length, offset + 1)) {
       // Consume an escaped code point. Append the returned code point to result.
       offset = consume_escaped(source, offset) - 1;
       offset += 1;
@@ -121,13 +121,13 @@ pub fn consume_number(source: &[u16], offset: usize) -> usize {
       code = source[offset];
 
       // 3. While the next input code point is a digit, consume it and append it to repr.
-      if is_digit!(code) {
+      if is_digit(code) {
         offset = find_decimal_number_end(source, offset + 1);
       }
       if offset + 1 < source_length {
         code = source[offset];
         // 4. If the next 2 input code points are U+002E FULL STOP (.) followed by a digit, then:
-        if code == 0x002E && is_digit!(source[offset + 1]) {
+        if code == 0x002E && is_digit(source[offset + 1]) {
           // 4.1 Consume them.
           // 4.2 Append them to repr.
           offset += 2;
@@ -142,7 +142,7 @@ pub fn consume_number(source: &[u16], offset: usize) -> usize {
       }
     }
 
-    if cmp_char!(source, source_length, offset, 101_u16 /* e */) != 0 {
+    if cmp_char(source, source_length, offset, 101_u16 /* e */) != 0 {
       let mut sign: usize = 0;
       if offset + 1 < source_length {
         code = source[offset + 1];
@@ -158,7 +158,7 @@ pub fn consume_number(source: &[u16], offset: usize) -> usize {
         }
 
         // ... followed by a digit
-        if !is_nan && is_digit!(code) {
+        if !is_nan && is_digit(code) {
           // 5.1 Consume them.
           // 5.2 Append them to repr.
 
@@ -190,7 +190,7 @@ pub fn consume_bad_url_remnants(source: &[u16], offset: usize) -> usize {
       return offset + 1;
     }
 
-    if is_valid_escape!(code, get_char_code!(source, source_length, offset + 1)) {
+    if is_valid_escape(code, get_char_code(source, source_length, offset + 1)) {
       // Consume an escaped code point.
       // Note: This allows an escaped right parenthesis ("\)") to be encountered
       // without ending the <bad-url-token>. This is otherwise identical to